feat: add URL Service functions
The 2 service functions for the URL are save_url and get_original_url.

save_url - It creates the record inside the URL table and saves it to the db and returns the new_url which is the shortened url.

get_original_url - It recieves either the original_url as parameter or the shortened_url as a paramater and which ever is entered gets queried on and retrieved.